get_info/requirements.txt
pandas>=2.0.0
google-generativeai>=0.8.0
tenacity>=8.2.0
